third_party/blink/web_tests/fast/dom/StyleSheet/css-medialist-item-expected.txt

This tests indexing outside the range of the media list object.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS mediaList.length is 2
PASS mediaList[0] is "screen"
PASS mediaList[1] is "print"
PASS mediaList[2] is undefined.
PASS mediaList[-1] is undefined.
PASS successfullyParsed is true

TEST COMPLETE
